+++ coverage.json
@@ -1,1 +1,1 @@
-{"meta": {"version": "7.2.1", "timestamp": "2023-03-10T13:12:44.611573", "branch_coverage": false, "show_contexts": false}, "files": {"uniswap_universal_router_decoder/__init__.py": {"executed_lines": [1], "summary": {"covered_lines": 0, "num_statements": 0, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/router_decoder.py": {"executed_lines": [1, 8, 10, 14, 15, 16, 17, 29, 30, 31, 35, 39, 40, 41, 42, 51, 52, 53, 55, 56, 59, 60, 61, 62, 63, 64, 65, 67, 68, 70, 77, 78, 79, 80, 81, 83, 84, 85, 86, 87, 88, 89, 90, 91, 93, 102, 103, 104, 105, 106, 108, 109, 118, 119, 120, 121, 122, 123, 124, 131, 133, 134, 136, 138, 139, 140, 141, 142, 143, 144, 145, 146, 147, 148, 149, 150, 152, 153, 154, 156, 157, 158, 159, 160, 162, 169, 170, 171, 172, 173, 175, 182, 183, 184, 185, 186, 188, 195, 196, 197, 198, 199, 200, 202, 209, 210, 211, 212, 213, 214, 216, 220, 225, 226, 227, 228, 229, 231, 232, 236, 238, 239, 243, 245, 246, 247, 248, 250, 259, 260, 267, 269, 286, 287, 288, 295, 302, 304, 320, 321, 322, 329, 336, 338, 356, 357, 358, 365, 372, 374, 391, 392, 393, 400, 407, 409, 410, 440, 446, 451, 452, 453, 454, 456, 469, 470, 477, 479, 500, 501, 502, 503, 510, 518, 520, 521, 522, 523, 524, 525, 527, 528, 530, 531, 541, 543, 544, 545, 546, 547, 549, 550, 551, 552, 553, 555, 556, 557, 558, 559, 560, 561, 562, 564, 565, 566, 567, 569, 570, 571, 572, 573, 575, 576, 577, 578, 580, 581, 582, 583, 584, 586, 587, 588, 589, 590, 593, 594, 595, 596, 597, 599, 600, 601, 602, 603, 605, 606, 609, 610, 611, 612, 615, 616, 617, 618, 619, 620, 621, 622, 625, 628, 629, 630, 632, 633, 634, 636, 637, 638, 640, 641, 642, 644, 645, 646, 648, 649, 651, 652, 653, 655, 656, 657, 659, 660, 661, 664, 666], "summary": {"covered_lines": 278, "num_statements": 278, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}}, "totals": {"covered_lines": 278, "num_statements": 278, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}}
+{"meta": {"version": "7.2.1", "timestamp": "2023-03-15T11:41:03.994958", "branch_coverage": false, "show_contexts": false}, "files": {"uniswap_universal_router_decoder/__init__.py": {"executed_lines": [1], "summary": {"covered_lines": 0, "num_statements": 0, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/_abi_builder.py": {"executed_lines": [1, 8, 10, 14, 21, 23, 26, 27, 28, 29, 30, 32, 33, 34, 35, 36, 38, 39, 42, 43, 44, 45, 48, 51, 52, 53, 55, 56, 57, 59, 60, 61, 63, 64, 65, 67, 68, 69, 71, 72, 74, 75, 76, 78, 79, 80, 82, 83, 84, 87, 88, 89, 99, 101, 102, 103, 104, 105, 107, 108, 109, 110, 111, 113, 114, 115, 116, 117, 118, 119, 120, 122, 123, 124, 125, 127, 128, 129, 130, 131, 133, 134, 135, 136, 138, 139, 140, 141, 142, 144, 145, 146, 147, 148], "summary": {"covered_lines": 93, "num_statements": 93, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/_constants.py": {"executed_lines": [1, 8, 13, 16, 17, 18, 20], "summary": {"covered_lines": 6, "num_statements": 6, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/_decoder.py": {"executed_lines": [1, 8, 9, 17, 18, 19, 26, 27, 28, 31, 32, 33, 34, 35, 37, 44, 45, 46, 47, 48, 50, 51, 52, 53, 54, 55, 56, 57, 58, 60, 69, 70, 71, 72, 73, 75, 76, 78, 79, 88, 89, 90, 91, 92, 93, 94, 101, 103, 104, 106], "summary": {"covered_lines": 49, "num_statements": 49, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/_encoder.py": {"executed_lines": [1, 8, 9, 20, 21, 22, 23, 24, 25, 26, 32, 33, 38, 41, 42, 43, 44, 45, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 61, 62, 63, 65, 66, 67, 68, 69, 71, 78, 79, 80, 81, 82, 84, 91, 92, 93, 94, 95, 97, 104, 105, 106, 107, 108, 109, 111, 118, 119, 120, 121, 122, 123, 125, 129, 134, 135, 136, 137, 138, 140, 141, 145, 147, 148, 152, 154, 155, 156, 157, 159, 168, 169, 176, 178, 195, 196, 197, 204, 211, 213, 229, 230, 231, 238, 245, 247, 265, 266, 267, 274, 281, 283, 300, 301, 302, 309, 316, 318, 331, 332, 339, 341, 362, 363, 364, 365, 372, 380, 382, 383, 384, 385, 386, 387], "summary": {"covered_lines": 126, "num_statements": 126, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/_enums.py": {"executed_lines": [1, 10, 13, 15, 16, 17, 18, 19, 20, 21], "summary": {"covered_lines": 9, "num_statements": 9, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}, "uniswap_universal_router_decoder/router_codec.py": {"executed_lines": [1, 9, 16, 20, 21, 26, 27, 28, 29, 32, 33, 34, 37, 38, 39, 40, 41, 42, 43, 45, 46, 47, 49, 50, 80, 86, 91, 92, 93, 94], "summary": {"covered_lines": 29, "num_statements": 29, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}, "missing_lines": [], "excluded_lines": []}}, "totals": {"covered_lines": 312, "num_statements": 312, "percent_covered": 100.0, "percent_covered_display": "100", "missing_lines": 0, "excluded_lines": 0}}
