+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "uniswap-universal-router-decoder"
-version = "0.7.1"
+version = "0.8.0.dev0"
 authors = [
   { name="Elnaril", email="elnaril_dev@caramail.com" },
 ]
