+++ tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from uniswap_universal_router_decoder.router_codec import RouterCodec
+
+
+@pytest.fixture
+def codec():
+    return RouterCodec()
