+++ uniswap_universal_router_decoder/_enums.py
@@ -1,0 +1,21 @@
+"""
+Enums used by the Uniswap Universal Router Codec
+
+* Author: Elnaril (https://www.fiverr.com/elnaril, https://github.com/Elnaril).
+* License: MIT.
+* Doc: https://github.com/Elnaril/uniswap-universal-router-decoder
+"""
+
+
+from enum import Enum
+
+
+class _RouterFunction(Enum):
+    # https://docs.uniswap.org/contracts/universal-router/technical-reference#command
+    V3_SWAP_EXACT_IN = 0
+    V3_SWAP_EXACT_OUT = 1
+    V2_SWAP_EXACT_IN = 8
+    V2_SWAP_EXACT_OUT = 9
+    PERMIT2_PERMIT = 10
+    WRAP_ETH = 11
+    UNWRAP_WETH = 12
